+++ libs/ImpSpec/test/Main.hs
@@ -1,0 +1,7 @@
+module Main where
+
+import Test.ImpSpec
+import Test.Suite.ImpSpec (spec)
+
+main :: IO ()
+main = impSpecMain $ describe "ImpSpec" spec